## Summary
- Implements complete gameplay voucher system with 13 vouchers affecting core game mechanics
- Extends VoucherEffect system with gameplay-specific effects (hand size, economy, penalties)
- Creates comprehensive voucher implementations.rs with trait-based architecture
- Integrates with game state modification system for permanent gameplay changes

## Gameplay Vouchers Implemented
- **Hand Management**: Grabber, NachoTong, Wasteful (+1 hand size variations)
- **Economic System**: SeedMoney, MoneyTree (interest cap modifications)
- **Strategic Penalties**: Hieroglyph, Petroglyph (extended game with tradeoffs)
- **Utility Enhancements**: Antimatter (+1 joker slot), Magic Trick/Illusion (shop cards)
- **Consumable Frequency**: TarotMerchant, TarotTycoon (2X/4X tarot frequency)
- **Specialty**: Blank (flavor), PaintBrush (mixed effects)

## Technical Architecture
- Extended VoucherEffect enum with 9 new effect types including negative effects
- Created modular implementations.rs with 13 complete voucher classes
- Proper prerequisite chains (MoneyTree requires SeedMoney, TarotTycoon requires TarotMerchant)
- Comprehensive validation and bounds checking for all new effects

## Dependencies
**Depends on**: PR #706 (Issue #17 shop voucher action system)

🤖 Generated with [Claude Code](https://claude.ai/code)